requirements.txt
libs/dlib-19.24.1-cp311-cp311-win_amd64.whl
opencv-python==4.7.0.72
face-recognition==1.3.0
numpy==1.26.0
Pillow==10.0.0
imutils==0.5.4
flask==2.3.2